add chatroom & message resources

+++ apps/api/src/message/message.service.ts
@@ -2,6 +2,7 @@ import {HttpException, HttpStatus, Injectable} from '@nestjs/common';
 import {CreateMessageInput} from './dto/create-message.input';
 import {UpdateMessageInput} from './dto/update-message.input';
 import {PrismaService} from "../prisma/prisma.service";
+import {LoggedUser} from "../shared/interfaces";
 
 @Injectable()
 export class MessageService {
@@ -29,10 +30,13 @@ export class MessageService {
     return await this.prisma.message.update({where: {id}, data: updateMessageInput});
   }
 
-  async remove(id: string) {
+  async remove(id: string, loggedUser: LoggedUser) {
     const message = await this.prisma.message.findUnique({where: {id}});
     if (!message) {
       throw new HttpException('Message not found', HttpStatus.BAD_REQUEST);
+    }
+    if(loggedUser.sub !== message.from || loggedUser.role !== 'admin') {
+      return new HttpException('You are not allowed to remove this message', HttpStatus.FORBIDDEN);
     }
     return await this.prisma.message.delete({where: {id}});
   }
